Start garbage collection after Infosquito reindex
I've observed that Infosquito holds on to a decent amount of garbage
after it's finished reindexing in our environments, and since it's then
sitting there doing nothing for a week that garbage tends to stick
around. Basically, since we have a long-running process that's very
active for a bit until it's suddenly inactive for a long time, we're
doing something strange with respect to the expectations of the Java GC.
So we may as well help it along a little.